Add files via upload
Create a directory for each of the videos that you want to process. Read .h5 files obtained using DeepLabCut from all the videos in a directory. Use the coordinates of the body parts and frame those body parts. Save the .png of only the framed parts and the frame with the squared parts so that you can see easily exactly what it's saved.